Removed NBT-purging logic for empty inventories
Previously, I overlooked the possibility of custom entries in the `BlockEntityTag`.
I had been consistently removing the entire parent object, relying solely on the content of its single child entry, namely `Items`.

Admittedly, this approach was flawed from the start.
Essentially, the mod tried to be helpful, but in reality, doing nothing was the best course of action.

Fixes #7
Fixes RubixDev/EnchantedShulkers#6